requirements.txt
Django~=2.2.4
django-markdownx==2.0.28